+++ spark/src/test/scala/org/apache/comet/parquet/ParquetReadSuite.scala
@@ -50,6 +50,7 @@ import com.google.common.primitives.UnsignedLong
 import org.apache.comet.CometConf
 import org.apache.comet.CometConf.SCAN_NATIVE_ICEBERG_COMPAT
 import org.apache.comet.CometSparkSessionExtensions.isSpark40Plus
+import org.apache.comet.rules.CometScanTypeChecker
 
 abstract class ParquetReadSuite extends CometTestBase {
   import testImplicits._
@@ -116,7 +117,9 @@ abstract class ParquetReadSuite extends CometTestBase {
       MapType(keyType = IntegerType, valueType = BinaryType) -> usingNativeIcebergCompat)
       .foreach { case (dt, expected) =>
         val fallbackReasons = new ListBuffer[String]()
-        assert(CometScanExec.isTypeSupported(dt, "", fallbackReasons) == expected)
+        assert(
+          CometScanTypeChecker(CometConf.COMET_NATIVE_SCAN_IMPL.get())
+            .isTypeSupported(dt, "", fallbackReasons) == expected)
         // usingDataFusionParquetExec does not support CometBatchScanExec yet
         if (!usingDataSourceExec(conf)) {
           assert(CometBatchScanExec.isTypeSupported(dt, "", fallbackReasons) == expected)
@@ -140,7 +143,9 @@ abstract class ParquetReadSuite extends CometTestBase {
     val fallbackReasons = new ListBuffer[String]()
 
     schemaDDLs.zip(cometScanExecSupported).foreach { case (schema, expected) =>
-      assert(CometScanExec.isSchemaSupported(StructType(schema), fallbackReasons) == expected)
+      assert(
+        CometScanTypeChecker(CometConf.COMET_NATIVE_SCAN_IMPL.get())
+          .isSchemaSupported(StructType(schema), fallbackReasons) == expected)
     }
 
     schemaDDLs.zip(cometBatchScanExecSupported).foreach { case (schema, expected) =>
